requirements.txt
Django==5.0.6
djangorestframework==3.15.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
